refactor(cli): flatten agent command module
