+++ turbo/apps/cli/src/commands/agent/avatar.ts
