+++ turbo/apps/cli/src/commands/agent/index.ts
@@ -5,7 +5,7 @@ import { viewCommand } from "./view";
 import { listCommand } from "./list";
 import { deleteCommand } from "./delete";
 
-export const zeroAgentCommand = new Command("agent")
+export const agentCommand = new Command("agent")
   .description("View or manage agents")
   .addCommand(createCommand)
   .addCommand(editCommand)
