+++ turbo/apps/cli/src/commands/agent/visibility.ts
